PRI-110: Refactor eligibility cookie for screeners
