+++ app/steps/ui/applicant/newexecutor/index.js
@@ -1,39 +1,26 @@
 'use strict';
 
-const ValidationStep = require('app/core/steps/ValidationStep');
+const EligibilityValidationStep = require('app/core/steps/EligibilityValidationStep');
 const content = require('app/resources/en/translation/applicant/newexecutor');
-const EligibilityCookie = require('app/utils/EligibilityCookie');
-const eligibilityCookie = new EligibilityCookie();
 const pageUrl = '/new-applicant-executor';
 const fieldKey = 'executor';
 
-class NewApplicantExecutor extends ValidationStep {
+class NewApplicantExecutor extends EligibilityValidationStep {
 
     static getUrl() {
         return pageUrl;
     }
 
     getContextData(req, res) {
-        const ctx = super.getContextData(req);
-
-        if (req.method === 'GET') {
-            const answerValue = eligibilityCookie.getAnswer(req, pageUrl, fieldKey);
-
-            if (answerValue) {
-                ctx[fieldKey] = answerValue;
-            }
-        } else {
-            const nextStepUrl = this.nextStepUrl(ctx);
-
-            this.setEligibilityCookie(req, res, nextStepUrl, fieldKey, ctx[fieldKey]);
-        }
-
-        return ctx;
+        return super.getContextData(req, res, pageUrl, fieldKey);
     }
 
     handlePost(ctx, errors, formdata, session) {
-        delete session.form;
-        return [ctx, errors];
+        return super.handlePost(ctx, errors, formdata, session);
+    }
+
+    persistFormData() {
+        return super.persistFormData();
     }
 
     nextStepUrl(ctx) {
@@ -46,14 +33,6 @@ class NewApplicantExecutor extends ValidationStep {
                 {key: fieldKey, value: content.optionYes, choice: 'isExecutor'}
             ]
         };
-    }
-
-    persistFormData() {
-        return {};
-    }
-
-    setEligibilityCookie(req, res, nextStepUrl, fieldKey, fieldValue) {
-        eligibilityCookie.setCookie(req, res, nextStepUrl, fieldKey, fieldValue);
     }
 }
