+++ test/unit/testNewIhtCompleted.js
@@ -3,10 +3,6 @@
 const initSteps = require('app/core/initSteps');
 const {expect} = require('chai');
 const content = require('app/resources/en/translation/iht/newcompleted');
-const rewire = require('rewire');
-const sinon = require('sinon');
-const schema = require('app/steps/ui/iht/newcompleted/schema');
-const newIhtCompleted = rewire('app/steps/ui/iht/newcompleted/index');
 const steps = initSteps([`${__dirname}/../../app/steps/action/`, `${__dirname}/../../app/steps/ui`]);
 const NewIhtCompleted = steps.NewIhtCompleted;
 
@@ -15,32 +11,6 @@ describe('NewIhtCompleted', () => {
         it('should return the correct url', (done) => {
             const url = NewIhtCompleted.constructor.getUrl();
             expect(url).to.equal('/new-iht-completed');
-            done();
-        });
-    });
-
-    describe('getContextData()', () => {
-        it('should call eligibilityCookie.getAnswer() with the correct params', (done) => {
-            const revert = newIhtCompleted.__set__('eligibilityCookie', {getAnswer: sinon.spy()});
-            const req = {method: 'GET', session: {form: {}}};
-            const res = {};
-
-            const steps = {};
-            const section = null;
-            const resourcePath = 'iht/newcompleted';
-            const i18next = {};
-            const newDeathCert = new newIhtCompleted(steps, section, resourcePath, i18next, schema);
-
-            newDeathCert.getContextData(req, res);
-
-            expect(newIhtCompleted.__get__('eligibilityCookie.getAnswer').calledOnce).to.equal(true);
-            expect(newIhtCompleted.__get__('eligibilityCookie.getAnswer').calledWith(
-                {method: 'GET', session: {form: {}}},
-                '/new-iht-completed',
-                'completed'
-            )).to.equal(true);
-
-            revert();
             done();
         });
     });
@@ -71,43 +41,6 @@ describe('NewIhtCompleted', () => {
                     choice: 'completed'
                 }]
             });
-            done();
-        });
-    });
-
-    describe('persistFormData()', () => {
-        it('should return an empty object', () => {
-            const result = NewIhtCompleted.persistFormData();
-            expect(result).to.deep.equal({});
-        });
-    });
-
-    describe('setEligibilityCookie()', () => {
-        it('should call eligibilityCookie.setCookie() with the correct params', (done) => {
-            const revert = newIhtCompleted.__set__('eligibilityCookie', {setCookie: sinon.spy()});
-            const req = {reqParam: 'req value'};
-            const res = {resParam: 'res value'};
-            const nextStepUrl = '/stop-page/ihtNotCompleted';
-            const fieldKey = 'completed';
-            const fieldValue = 'Yes';
-            const steps = {};
-            const section = null;
-            const resourcePath = 'iht/newcompleted';
-            const i18next = {};
-            const newIhtCom = new newIhtCompleted(steps, section, resourcePath, i18next, schema);
-
-            newIhtCom.setEligibilityCookie(req, res, nextStepUrl, fieldKey, fieldValue);
-
-            expect(newIhtCompleted.__get__('eligibilityCookie.setCookie').calledOnce).to.equal(true);
-            expect(newIhtCompleted.__get__('eligibilityCookie.setCookie').calledWith(
-                {reqParam: 'req value'},
-                {resParam: 'res value'},
-                '/stop-page/ihtNotCompleted',
-                'completed',
-                'Yes'
-            )).to.equal(true);
-
-            revert();
             done();
         });
     });
