+++ test/unit/testNewMentalCapacity.js
@@ -3,10 +3,6 @@
 const initSteps = require('app/core/initSteps');
 const {expect} = require('chai');
 const content = require('app/resources/en/translation/executors/newmentalcapacity');
-const rewire = require('rewire');
-const sinon = require('sinon');
-const schema = require('app/steps/ui/executors/newmentalcapacity/schema');
-const newMentalCapacity = rewire('app/steps/ui/executors/newmentalcapacity/index');
 const steps = initSteps([`${__dirname}/../../app/steps/action/`, `${__dirname}/../../app/steps/ui`]);
 const NewMentalCapacity = steps.NewMentalCapacity;
 
@@ -15,32 +11,6 @@ describe('NewMentalCapacity', () => {
         it('should return the correct url', (done) => {
             const url = NewMentalCapacity.constructor.getUrl();
             expect(url).to.equal('/new-mental-capacity');
-            done();
-        });
-    });
-
-    describe('getContextData()', () => {
-        it('should call eligibilityCookie.getAnswer() with the correct params', (done) => {
-            const revert = newMentalCapacity.__set__('eligibilityCookie', {getAnswer: sinon.spy()});
-            const req = {method: 'GET', session: {form: {}}};
-            const res = {};
-
-            const steps = {};
-            const section = null;
-            const resourcePath = 'executors/newmentalcapacity';
-            const i18next = {};
-            const newDeathCert = new newMentalCapacity(steps, section, resourcePath, i18next, schema);
-
-            newDeathCert.getContextData(req, res);
-
-            expect(newMentalCapacity.__get__('eligibilityCookie.getAnswer').calledOnce).to.equal(true);
-            expect(newMentalCapacity.__get__('eligibilityCookie.getAnswer').calledWith(
-                {method: 'GET', session: {form: {}}},
-                '/new-mental-capacity',
-                'mentalCapacity'
-            )).to.equal(true);
-
-            revert();
             done();
         });
     });
@@ -71,43 +41,6 @@ describe('NewMentalCapacity', () => {
                     choice: 'isCapable'
                 }]
             });
-            done();
-        });
-    });
-
-    describe('persistFormData()', () => {
-        it('should return an empty object', () => {
-            const result = NewMentalCapacity.persistFormData();
-            expect(result).to.deep.equal({});
-        });
-    });
-
-    describe('setEligibilityCookie()', () => {
-        it('should call eligibilityCookie.setCookie() with the correct params', (done) => {
-            const revert = newMentalCapacity.__set__('eligibilityCookie', {setCookie: sinon.spy()});
-            const req = {reqParam: 'req value'};
-            const res = {resParam: 'res value'};
-            const nextStepUrl = '/stop-page/mentalCapacity';
-            const fieldKey = 'mentalCapacity';
-            const fieldValue = 'Yes';
-            const steps = {};
-            const section = null;
-            const resourcePath = 'executors/newmentalcapacity';
-            const i18next = {};
-            const newMenCap = new newMentalCapacity(steps, section, resourcePath, i18next, schema);
-
-            newMenCap.setEligibilityCookie(req, res, nextStepUrl, fieldKey, fieldValue);
-
-            expect(newMentalCapacity.__get__('eligibilityCookie.setCookie').calledOnce).to.equal(true);
-            expect(newMentalCapacity.__get__('eligibilityCookie.setCookie').calledWith(
-                {reqParam: 'req value'},
-                {resParam: 'res value'},
-                '/stop-page/mentalCapacity',
-                'mentalCapacity',
-                'Yes'
-            )).to.equal(true);
-
-            revert();
             done();
         });
     });
