+++ test/unit/testNewWillOriginal.js
@@ -3,10 +3,6 @@
 const initSteps = require('app/core/initSteps');
 const {expect} = require('chai');
 const content = require('app/resources/en/translation/will/neworiginal');
-const rewire = require('rewire');
-const sinon = require('sinon');
-const schema = require('app/steps/ui/will/neworiginal/schema');
-const newWillOriginal = rewire('app/steps/ui/will/neworiginal/index');
 const steps = initSteps([`${__dirname}/../../app/steps/action/`, `${__dirname}/../../app/steps/ui`]);
 const NewWillOriginal = steps.NewWillOriginal;
 
@@ -15,32 +11,6 @@ describe('NewWillOriginal', () => {
         it('should return the correct url', (done) => {
             const url = NewWillOriginal.constructor.getUrl();
             expect(url).to.equal('/new-will-original');
-            done();
-        });
-    });
-
-    describe('getContextData()', () => {
-        it('should call eligibilityCookie.getAnswer() with the correct params', (done) => {
-            const revert = newWillOriginal.__set__('eligibilityCookie', {getAnswer: sinon.spy()});
-            const req = {method: 'GET', session: {form: {}}};
-            const res = {};
-
-            const steps = {};
-            const section = null;
-            const resourcePath = 'will/neworiginal';
-            const i18next = {};
-            const newDeathCert = new newWillOriginal(steps, section, resourcePath, i18next, schema);
-
-            newDeathCert.getContextData(req, res);
-
-            expect(newWillOriginal.__get__('eligibilityCookie.getAnswer').calledOnce).to.equal(true);
-            expect(newWillOriginal.__get__('eligibilityCookie.getAnswer').calledWith(
-                {method: 'GET', session: {form: {}}},
-                '/new-will-original',
-                'original'
-            )).to.equal(true);
-
-            revert();
             done();
         });
     });
@@ -71,43 +41,6 @@ describe('NewWillOriginal', () => {
                     choice: 'isOriginal'
                 }]
             });
-            done();
-        });
-    });
-
-    describe('persistFormData()', () => {
-        it('should return an empty object', () => {
-            const result = NewWillOriginal.persistFormData();
-            expect(result).to.deep.equal({});
-        });
-    });
-
-    describe('setEligibilityCookie()', () => {
-        it('should call eligibilityCookie.setCookie() with the correct params', (done) => {
-            const revert = newWillOriginal.__set__('eligibilityCookie', {setCookie: sinon.spy()});
-            const req = {reqParam: 'req value'};
-            const res = {resParam: 'res value'};
-            const nextStepUrl = '/stop-page/notOriginal';
-            const fieldKey = 'original';
-            const fieldValue = 'Yes';
-            const steps = {};
-            const section = null;
-            const resourcePath = 'will/neworiginal';
-            const i18next = {};
-            const newWilOri = new newWillOriginal(steps, section, resourcePath, i18next, schema);
-
-            newWilOri.setEligibilityCookie(req, res, nextStepUrl, fieldKey, fieldValue);
-
-            expect(newWillOriginal.__get__('eligibilityCookie.setCookie').calledOnce).to.equal(true);
-            expect(newWillOriginal.__get__('eligibilityCookie.setCookie').calledWith(
-                {reqParam: 'req value'},
-                {resParam: 'res value'},
-                '/stop-page/notOriginal',
-                'original',
-                'Yes'
-            )).to.equal(true);
-
-            revert();
             done();
         });
     });
